Rename files:  - dfc_pkg/deploy/dpl/nginx.p to dfc_pkg/deploy/nginx_server/setup_nginx.py  - dfc_pkg/deploy/nginx_project.py to  dfc_pkg/deploy/nginx.py This to re-organiced the project.
