+++ dfc_pkg/deploy/commands/verify_path.py
