+++ dfc_pkg/deploy/nginx.py
@@ -1,13 +1,14 @@
 #!/usr/bin/python3
-from deploy.dpl.verify_path import verify_path
-from deploy.dpl.compress import compress
-from deploy.dpl.create_connection import create_connection
-from deploy.dpl.nginx import install_nginx, upload_files, config_nginx, run_gunicorn
+from deploy.commands.verify_path import verify_path
+from deploy.commands.compress import compress
+from deploy.commands.create_connection import create_connection
+from deploy.commands.upload_files import upload_files
+from deploy.nginx_server.setup_nginx import install_nginx, config_nginx, run_gunicorn
 import os
 import typer
 
 app = typer.Typer()
-path_proyect = os.getenv("DFC_PATH")
+path_project = os.getenv("DFC_PATH")
 
 
 @app.command()
@@ -21,22 +22,22 @@ def nginx_project(ip: str = typer.Option(...), path_key: str = typer.Option(...)
     :param port: the number of the port to run gunicorn
     :return: deploy server
     """
-    if path_proyect is None:
+    if path_project is None:
         os.chdir(os.path.dirname(__file__))
         verify_path(os.getcwd())
         zip_file = compress(os.getcwd())
         server = create_connection(user_ssh, ip, path_key)
         install_nginx(server, user_ssh, ip)
-        upload_files(os.getcwd(), server, zip_file)
+        upload_files(server, os.getcwd(), zip_file)
+        config_nginx(server, os.getcwd(), zip_file, user_ssh)
     else:
-        verify_path(path_proyect)
-        zip_file = compress(path_proyect)
+        verify_path(path_project)
+        zip_file = compress(path_project)
         server = create_connection(user_ssh, ip, path_key)
         install_nginx(server, user_ssh, ip)
-        upload_files(server, path_proyect, zip_file)
-        config_nginx(server, path_proyect, zip_file, user_ssh)
-        run_gunicorn(server, path_proyect, port)
-
+        upload_files(server, path_project, zip_file)
+        config_nginx(server, path_project, zip_file, user_ssh)
+        # run_gunicorn(server, path_project, port)
 
 
 if __name__ == "__main__":
