+++ dfc_pkg/main.py
@@ -1,10 +1,18 @@
 #!/usr/bin/python3
+"""
+    # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+    #  Command line interpreter for Deploy Flask CLI (DFC)  #
+    # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+    :deploy:
+        Create the deploy for NGINX server v0.1
+"""
 import typer
-import deploy.nginx_project as nginx_project
+import deploy.nginx as nginx
 
 app = typer.Typer()
 
-app.add_typer(nginx_project.app, name="deploy")
+app.add_typer(nginx.app, name="deploy")
 
 if __name__ == "__main__":
     app()
